+++ packages/core_media_data/lib/core_media_data.dart
@@ -1,4 +1,5 @@
 library;
 
 export "src/media_database_benchmark_models.dart";
+export "src/media_graph.dart";
 export "src/profile_data_ownership_models.dart";